ethos-drive/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ethos-drive"
version = "1.0.0"
description = "Desktop sync client for EthOS NAS"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}

dependencies = [
    "PySide6>=6.6.0",
    "httpx>=0.27.0",
    "python-socketio[client]>=5.11.0",
    "watchdog>=4.0.0",
    "xxhash>=3.4.0",
    "keyring>=25.0.0",
    "platformdirs>=4.2.0",
    "pydantic>=2.6.0",
    "certifi>=2024.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-qt>=4.4.0",
    "pyinstaller>=6.4.0",
]

[project.scripts]
ethos-drive = "ethos_drive.main:main"

[project.gui-scripts]
ethos-drive-gui = "ethos_drive.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
